Rename file to match class
