+++ Library/Homebrew/extend/git_path.rb
