+++ Library/Homebrew/extend/git_path.rbi
