+++ Library/Homebrew/global.rb
@@ -128,7 +128,7 @@ end
 
 require "context"
 require "extend/array"
-require "extend/git_repository"
+require "extend/git_path"
 require "extend/pathname"
 require "extend/predicable"
 require "extend/module"
